Поле для выбора величины рейза и логика отрисовки вариантов выбора

+++ PocketPoker/core/src/game/poker/gui/table/SeatHorizontal.kt
@@ -1,6 +1,6 @@
 package game.poker.gui.table
 
-import java.awt.Point
+import game.poker.gui.table.TableViewBase.Point
 
 class SeatHorizontal(private val positionNumber:Int, private val potPosition: Point) : SeatBase(){
     override fun updateCardsPosition(){
@@ -27,56 +27,56 @@ class SeatHorizontal(private val positionNumber:Int, private val potPosition: Po
         when(positionNumber){
             0 -> {
                 setPosition(120f, 1100f)
-                chipstackPosition.setLocation(350, 150)
+                chipstackPosition.setLocation(350f, 150f)
                 dealerChip.setPosition(400f, 30f)
                 playerView.setPosition(-270f, 0f)
             }
             1 -> {
                 setPosition(140f, 1800f)
-                chipstackPosition.setLocation(250, 200)
+                chipstackPosition.setLocation(250f, 200f)
                 dealerChip.setPosition(180f,200f)
                 playerView.setPosition(-100f, 150f)
             }
             2 -> {
                 setPosition(460f, 1900f)
-                chipstackPosition.setLocation(300, 60)
+                chipstackPosition.setLocation(300f, 60f)
                 dealerChip.setPosition(230f,60f)
                 playerView.setPosition(0f, 150f)
             }
             3 -> {
                 setPosition(780f, 1640f)
-                chipstackPosition.setLocation(100, -90)
+                chipstackPosition.setLocation(100f, -90f)
                 dealerChip.setPosition(30f,-140f)
                 playerView.setPosition(-260f,0f)
             }
             4 -> {
                 setPosition(900f, 1430f)
-                myPotPosition.setLocation(340, -500)
-                chipstackPosition.setLocation(50, -80)
+                myPotPosition.setLocation(340f, -500f)
+                chipstackPosition.setLocation(50f, -80f)
                 dealerChip.setPosition(270f,-100f)
                 playerView.setPosition(210f, 0f)
             }
             5 -> {
                 setPosition(900f, 970f)
-                chipstackPosition.setLocation(50, -80)
+                chipstackPosition.setLocation(50f, -80f)
                 dealerChip.setPosition(270f,-100f)
                 playerView.setPosition(210f, 0f)
             }
             6 -> {
                 setPosition(800f, 500f)
-                chipstackPosition.setLocation(-100, -90)
+                chipstackPosition.setLocation(-100f, -90f)
                 dealerChip.setPosition(100f,-90f)
                 playerView.setPosition(210f, 0f)
             }
             7 -> {
                 setPosition(460f, 250f)
-                chipstackPosition.setLocation(-260, 60)
+                chipstackPosition.setLocation(-260f, 60f)
                 dealerChip.setPosition(-60f,60f)
                 playerView.setPosition(-40f, 150f)
             }
             8 -> {
                 setPosition(140f, 400f)
-                chipstackPosition.setLocation(-200, 250)
+                chipstackPosition.setLocation(-200f, 250f)
                 dealerChip.setPosition(30f,200f)
                 playerView.setPosition(100f, 150f)
             }
@@ -89,8 +89,8 @@ class SeatHorizontal(private val positionNumber:Int, private val potPosition: Po
             card1.setSize(100f, 140f)
             card2.setSize(100f, 140f)
         }
-        myPotPosition.setLocation(y.toInt() - potPosition.y, potPosition.x - x.toInt())
-        chipstack.setPosition(chipstackPosition.x.toFloat(), chipstackPosition.y.toFloat())
+        myPotPosition.setLocation(y - potPosition.y, potPosition.x - x)
+        chipstack.setPosition(chipstackPosition.x, chipstackPosition.y)
         card1.setPosition(0f,0f)
         card2.setPosition(0f,0f)
         updateCardsPosition()
